+++ slack-sup/commands/mixins.rb
@@ -1,1 +1,2 @@
 require 'slack-sup/commands/mixins/subscribe'
+require 'slack-sup/commands/mixins/pluralize'