+++ slack-sup/commands/next.rb
@@ -5,14 +5,7 @@ module SlackSup
 
       subscribe_command 'next' do |client, data, _match|
         team = client.owner
-        next_sup_at = team.next_sup_at
-        messages = [
-          'Next round is',
-          Time.now > next_sup_at ? 'overdue' : nil,
-          next_sup_at.strftime('%A, %B %e, %Y at %l:%M %p %Z').gsub('  ', ' '),
-          '(' + next_sup_at.to_time.ago_or_future_in_words(highest_measure_only: true) + ').'
-        ].compact.join(' ')
-        client.say(channel: data.channel, text: messages)
+        client.say(channel: data.channel, text: team.next_sup_at_text)
         logger.info "NEXT: #{client.owner} - #{data.user}"
       end
     end